Add Int64 type parameterization back in

+++ calibration/calibration_helper_functions.jl
@@ -25,7 +25,7 @@ using CSVFiles
 #----------------------------------------------------------------------------------------------------------------------
 
 
-function load_calibration_data(model_start_year::Int, last_calibration_year; last_sea_level_norm_year::Int=1990)
+function load_calibration_data(model_start_year::Int64, last_calibration_year; last_sea_level_norm_year::Int=1990)
 
     # Create column of calibration years and calculate indicies for calibration time period relative to 1765-2020 (will crop later).
     # Note: first year is first year to run model (not necessarily year of first observation).
